Queen buff
Wave does twice as much base damage. It's easy enough to avoid so if you get hit by it sucks to be you. Players will definitely misunderstand the inverse thing but I didn't want to mess withit too much. Will buffs by 2 instead of 1 and defend also adds a Void to the player's draw pile. Hopefully that's not too overboard